+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 is now available in the ```dev-env``` docker image
 * added ```check-consistent-dev-env-version.sh``` to detect dev-env version drift
 * ```install-dev-env.sh``` is now the recommended way for ```dev-env``` to be installed by a ```cfg4dev``` script
+* added ```python-increment-version.sh``` for use in ```.cut-release-master-branch-changes.sh```
 
 ### Changed